通过登录成功再封装会话数据以供判断实现上一个版本保护登陆后页面的功能

+++ webone/src/com/learn/web/UserServlet.java
@@ -12,7 +12,7 @@ import javax.servlet.http.HttpSession;
 import java.io.IOException;
 import java.util.List;
 
-@WebServlet(name = "UserServlet",urlPatterns = {"/login/userServlet","/WEB-INF/LoginSuccess/UserServlet"})
+@WebServlet(name = "UserServlet",urlPatterns = "/login/userServlet")
 public class UserServlet extends HttpServlet {
     //请求的总控制中心
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
@@ -23,10 +23,7 @@ public class UserServlet extends HttpServlet {
             this.doAdd(request,response);
         }else if ("delete".equals(method)){
             this.doDelete(request,response);
-        }else if ("changePage".equals(method)){
-            this.getChangePage(request,response);
-        }
-        else if ("change".equals(method)){
+        } else if ("change".equals(method)){
             this.doChange(request,response);
         }
     }
@@ -42,8 +39,10 @@ public class UserServlet extends HttpServlet {
         UserDaoImpl dao=new UserDaoImpl();//构建一个用户表的数据访问对象
         List<User> users = dao.selectAll();//获取所有的用户列表
         if (users.contains(user)){
-            req.setAttribute("list",users);//封装数据到请求中
-            req.getRequestDispatcher("/WEB-INF/successLogin/Success.jsp").forward(req,resp);
+            HttpSession session = req.getSession();//获取会话
+            session.setAttribute("currentUser",user);//将当前登录的用户封装到会话中，用于后续网页的展示与判断（多页面展示）
+            req.setAttribute("list",users);//将所有的用户信息封装数据到请求中（单页面展示）
+            req.getRequestDispatcher("/login/Success.jsp").forward(req,resp);
         }else {//用户名和密码在数据库中不存在
             resp.setContentType("text/html;charset=UTF-8");
             resp.getWriter().print("<script>alert('用户名或密码错误');location.href='/login/Login.jsp'</script>");
@@ -64,9 +63,11 @@ public class UserServlet extends HttpServlet {
         int flag=dao.insert(user);
         if (flag>0){
             //插入成功，直接登录
+            HttpSession session = req.getSession();//获取会话
+            session.setAttribute("currentUser",user);//将当前登录的用户封装到会话中，用于后续网页的展示与判断（多页面展示）
             List<User> users = dao.selectAll();//获取全部数据
             req.setAttribute("list",users);//封装数据到请求中
-            req.getRequestDispatcher("/WEB-INF/successLogin/Success.jsp").forward(req,resp);
+            req.getRequestDispatcher("/login/Success.jsp").forward(req,resp);
         }else {//用户已经存在，继续留在登录页面
             resp.setContentType("text/html;charset=UTF-8");//设置返回内容的编码格式
             //给出提示，返回注册页面
@@ -77,22 +78,23 @@ public class UserServlet extends HttpServlet {
     }
     //删除服务中心
     protected void doDelete(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        //将前端的请求数据解码
-        req.setCharacterEncoding("utf-8");
-        String userName = req.getParameter("userName");//获取超链接提交的需要删除的用户名
-        userName=new String(userName.getBytes("iso-8859-1"),"UTF-8");
-        UserDaoImpl dao=new UserDaoImpl();
-        User user=new User(userName);
-        int delete = dao.delete(user);//删除成功返回1，否则为0
-        //然后将删除后的数据返回
-        //因为删除是针对确定已经有的数据进行，所以无需判断，直接返回结果
-        List<User> users = dao.selectAll();
-        req.setAttribute("list",users);//封装数据到请求中
-        req.getRequestDispatcher("/WEB-INF/successLogin/Success.jsp").forward(req,resp);
-    }
-    //获得更改数据的页面
-    protected void getChangePage(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        req.getRequestDispatcher("/WEB-INF/successLogin/UpdateUser.jsp").forward(req,resp);
+        //通过判断是否具有当前登录用户，可以过滤掉直接地址栏访问的请求
+        if (req.getSession().getAttribute("currentUser")!=null) {
+            //将前端的请求数据解码
+            req.setCharacterEncoding("utf-8");
+            String userName = req.getParameter("userName");//获取超链接提交的需要删除的用户名
+            userName=new String(userName.getBytes("iso-8859-1"),"UTF-8");
+            UserDaoImpl dao=new UserDaoImpl();
+            User user=new User(userName);
+            int delete = dao.delete(user);//删除成功返回1，否则为0
+            //然后将删除后的数据返回
+            //因为删除是针对确定已经有的数据进行，所以无需判断，直接返回结果
+            List<User> users = dao.selectAll();
+            req.setAttribute("list",users);//封装数据到请求中
+            req.getRequestDispatcher("/login/Success.jsp").forward(req,resp);
+        }else {//如果有人试图想不登录删除，我就让他去登录
+            resp.sendRedirect("/login/Login.jsp");
+        }
     }
     //更改服务中心
     protected void doChange(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
@@ -103,11 +105,11 @@ public class UserServlet extends HttpServlet {
         if (update>0){
             List<User> users = dao.selectAll();//获取当前数据库的所有信息
             req.setAttribute("list",users);//封装数据到请求中
-            req.getRequestDispatcher("/WEB-INF/successLogin/Success.jsp").forward(req,resp);
+            req.getRequestDispatcher("/login/Success.jsp").forward(req,resp);
         }else {
             resp.setContentType("text/html;charset=UTF-8");//设置回应数据的编码
             //提示错误信息，并返回更改页面
-            resp.getWriter().print("<script>alert('你要更改的用户不存在');location.href='/login/userServlet?method=changePage'</script>");
+            resp.getWriter().print("<script>alert('你要更改的用户不存在');location.href='/login/UpdateUser.jsp'</script>");
         }
     }
 }
